feat(Issue Detail View): added new IssuesCommentsFilter type
This commit adds a new type, IssuesCommentsFilter, to allow for sorting comments. The new type
follows the same style as the default Filter and could possibly be expanded on to filter comments by
author, or order the comments based on different values.
This commit also adds an
issuesCommentFilters key to the UiState type.