feat(Browsing Issues): added functionality to favorite projects
Users now have the ability to favorite projects! A star-icon is present next to the
IssueSourcePicker and there is also a new category called 'Favorite Projects' which contains a
section of all favorite projects for the user to scroll through. The section appears before the rest
to give users an easier time finding favorite projects.
Favorites are also persisted to electron
storage on quit and reloaded whenever the app is restarted.
This only works for projects at the
moments but can be used for filters or other issueSourceTypes later.

ISSUES CLOSED: #47